Type the camera fine-coordinate bearing/range and Q8 eye offsets
The tracked/tracking camera path is the last "camera precision" gap:
computeBearing32/rangeApprox32 on fine deltas and the sinMulQ8/cosMulQ8
eye offsets now run through GuidanceMath helpers:

- wideBearing: fixed runs the shared shift-to-fit + LUT bearing on
  int32-narrowed deltas; modern atan2 on un-narrowed fractional deltas.
- wideRange: rangeApprox32's max+min/2 at full width, fraction under modern.
- sineOffsetQ8/cosineOffsetQ8: (sine*mag)>>7 under fixed; trig*mag*2^8
  under modern.
- eyeFromQ8 takes the Q8 sum as double so fractional fine positions
  (fineRep) and trig products carry real sub-word precision into the
  renderer's frac byte; identical to >>8/&0xff on integer inputs.

drawTargetView's dxFine/dyFine are fraction-capable, and renderFrame's
external/follow/target camera eyes keep g_ourHead typed through the trig.
The deliberate float offset block (keyed to the view matrix's sine table)
and all word/int32 camera outputs stay at the render boundary.

Verification: fixed 59/59 (sortie parity exercises external/tracking
views), modern smoke, wideBearing/wideRange/Q8 helpers oracle-tested
against computeBearing32/rangeApprox32/sinMulQ8/cosMulQ8, analyzer
warnings all pre-existing.

+++ src/egflight.h
@@ -32,8 +32,8 @@ void brakeFlightSpeed();
 void advanceFlightHorizontal(f15::math::HorizontalSpeed<f15::math::GameBackend> speed);
 
 int16 isqrt(int16 value);
-void computeTrackingCameraAngles(int32 targetX, int32 targetY, int16 targetAlt,
-                                 int32 viewX, int32 viewY, int16 viewAlt,
+void computeTrackingCameraAngles(double targetX, double targetY, int16 targetAlt,
+                                 double viewX, double viewY, int16 viewAlt,
                                  int16 *heading, int16 *pitch);
 void UpdateThrottleState(void);
 void drawFuelGauge(void);
